# ns-topup/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NS Games Topup API",
        "description": "API untuk platform topup game, pulsa, dan e-wallet dengan integrasi Digiflazz dan Midtrans",
        "contact": {
            "name": "NS Games Support",
            "email": "admin@nsgames.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Development Server"
        },
        {
            "url": "http://localhost/NS-topupgames/ns-topup/public",
            "description": "Laragon/XAMPP Server"
        }
    ],
    "paths": {
...
        "/api/admin/games": {
            "get": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Get all games for admin",
                "operationId": "113497d56bcc3bc9e07b88b17057e91b",
                "responses": {
                    "200": {
                        "description": "List of all games",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Mobile Legends"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "MOBA"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "MOBA game populer"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Create new game",
                "operationId": "6665bc27602de5ae2f2cc9e9431fef1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mobile Legends"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "MOBA"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "MOBA game populer"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/games/{id}": {
            "put": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Update game",
                "operationId": "e8f7a714868f482f6b1a5a64feb8d73c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Delete game",
                "operationId": "1ec5823f5631feae31a4c5f492b43e37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}